// RMG/ClientApp/src/app/models/Employee.ts
export class Employee
{
  bg_id: string;
  emp_Id: string;
  emp_Name:string;
  designation_Id:string;
  department_Id:string;
  edge_Practice_Id:string;
  coe_Id:string;
  location_Id: string;
  location: string;
  joining_Date: string;
  contact_Number: string;
  address: string;
  email_Id:string;
  reporting_To: string;
  reporting_To_Email: string;
  job_Description: string;
  cat_Description: string;
  sub_Cat_Description: string;
  flag_Status: string;
  created_By: string;
  last_Updated_By: string;

}